Update Project1_pxm006.py
A project submitted as part of my Intro to Comp Programming class.
A python script which serves as a rudimentary student grade menu which uses the user's inputs to create a dictionary that contains a student's name and assigned grade. The user is then able to edit pre-existing student and grade information, view a list of students and their grades as well as a class grade report, or terminate the program using a choice input.