Python Project 2: Student Grade Report 2
An expanded project based on the previous submitted as part of my Intro to Comp Programming class.
A python script which serves as a rudimentary student grade menu which uses a separate .csv file to make a viewable grade report of student entries in said grade report. The script uses choice inputs to allow the user to either view a particular student's listed exam grades and averages, view a graph of a student's or the class's reported exam grades, and/or create a file of the class's average exam grades.